Application tier --> Front-end
I prefer the term application tier, as it's a more accurate description
of the function this role performs (or perhaps a more relaxed term, at
least).

However, it has been highlighted that the rest of the documentation
refers to front-ends.  Make this commit consistent with the rest of the
documentation, and look to change the rest of the documentation to the
new standard separately.