Only decrypt secrets when doing a deployment
Our pull request builds are failing because travis doesn't make our
secrets available to potential strangers.  This is a good thing, but
there's no need to decrypt these secrets if we're not intending to
deploy.

Put in some dodgy logic to not attempt decryption on builds we don't
intend to deploy.